Connection diagnostics
Display is toggled by clicking on the signal delay field in the Time Wrap display.
Connection is cached internally, and the last cached route is checked for quality.

Limitation: because no RT2 information is retained between scene changes, diagnostics
are only available for the active vessel, and only if the connection goes down while
the vessel is active.